In-band GAS read & GAS dump output sanity check
To distinguish all 1 returns of in-band GAS read & GAS dump valid
or not, a light way rely on check of device id is added.
Raise the SIGABRT signal when the output is abnormal.

For the reason of Windows system has no define of SIGBUS in
SIGNAL.H, SIGABRT is choosed.